ml_engine/requirements.txt
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.3.0
spacy==3.7.5
joblib==1.4.2
matplotlib==3.9.0
